Remove functions from edit.js
